backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
httpx==0.28.1
boto3==1.35.88
chromadb==0.5.23
python-multipart==0.0.19
